requirements.txt
ccxt==4.1.33
python-dotenv==0.20.0
finta==1.3
pandas==2.1.2
ta==0.10.3
numpy==1.26.1